remove code from zuora query client file
