+++ handlers/product-move-api/src/main/scala/com/gu/productmove/zuora/InvoiceItemQuery.scala
@@ -67,48 +67,6 @@ trait InvoiceItemQuery {
   def invoiceItemsForSubscription(subscriptionName: SubscriptionName): ZIO[Any, Throwable, List[InvoiceItem]]
   def taxationItemsForInvoice(invoiceId: InvoiceId): ZIO[Any, Throwable, List[TaxationItem]]
 }
-case class RefundInvoiceDetails(
-    refundAmount: BigDecimal,
-    negativeInvoiceId: InvoiceId,
-    negativeInvoiceItems: List[InvoiceItemWithTaxDetails],
-    lastPaidInvoiceId: InvoiceId,
-)
-case class TaxDetails(amount: BigDecimal, taxationId: String)
-
-object InvoiceItemWithTaxDetails {
-
-  def apply(i: InvoiceItem, taxationItems: List[InvoiceItemQuery.TaxationItem]): InvoiceItemWithTaxDetails = {
-    val maybeTaxDetails = for {
-      negativeInvoiceItem <- Some(i).filter(_.TaxAmount != 0)
-      _ = println(
-        s"Tax amount for invoice item $i is ${negativeInvoiceItem.TaxAmount} searching for matching taxation item",
-      )
-      taxationItem <- taxationItems.find(_.InvoiceItemId == negativeInvoiceItem.Id)
-      _ = println(s"Found taxation item $taxationItem")
-    } yield TaxDetails(i.TaxAmount, taxationItem.Id)
-    InvoiceItemWithTaxDetails(
-      Id = i.Id,
-      AppliedToInvoiceItemId = i.AppliedToInvoiceItemId,
-      ChargeDate = i.ChargeDate,
-      ChargeAmount = i.ChargeAmount,
-      TaxDetails = maybeTaxDetails,
-      InvoiceId = i.InvoiceId,
-    )
-  }
-
-}
-case class InvoiceItemWithTaxDetails(
-    Id: String,
-    AppliedToInvoiceItemId: Option[String] = None,
-    ChargeDate: String,
-    ChargeAmount: BigDecimal,
-    TaxDetails: Option[TaxDetails],
-    InvoiceId: InvoiceId,
-) {
-  def taxAmount = TaxDetails.map(_.amount).getOrElse(BigDecimal(0))
-  def amountWithTax = ChargeAmount + taxAmount
-  def chargeDateAsDate = LocalDate.parse(ChargeDate, DateTimeFormatter.ISO_OFFSET_DATE_TIME)
-}
 
 object InvoiceItemQuery {
