+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/GetRefundInvoiceDetailsLiveSpec.scala
@@ -1,10 +1,9 @@
 package com.gu.productmove.zuora
 
-import com.gu.productmove.{AwsCredentialsLive, GuStageLive, SQSLive, SecretsLive, SttpClientLive}
+import com.gu.productmove.*
 import com.gu.productmove.endpoint.cancel.SubscriptionCancelEndpoint
 import com.gu.productmove.endpoint.cancel.SubscriptionCancelEndpointTypes.ExpectedInput
-import com.gu.productmove.refund.{GetRefundInvoiceDetails, RefundSupporterPlus}
-import com.gu.productmove.zuora.TaxDetails
+import com.gu.productmove.refund.{GetRefundInvoiceDetails, RefundSupporterPlus, TaxDetails}
 import com.gu.productmove.zuora.model.SubscriptionName
 import com.gu.productmove.zuora.rest.{ZuoraClientLive, ZuoraGetLive}
 import zio.*
