+++ handlers/product-move-api/src/test/scala/com/gu/productmove/zuora/manual/InvoiceItemAdjustmentSpec.scala
@@ -9,16 +9,10 @@ import com.gu.productmove.invoicingapi.InvoicingApiRefundLive
 import com.gu.productmove.refund.*
 import com.gu.productmove.zuora.GetSubscription.GetSubscriptionResponse
 import com.gu.productmove.zuora.InvoiceItemAdjustment.InvoiceItemAdjustmentResult
-import com.gu.productmove.zuora.model.InvoiceId
 import com.gu.productmove.zuora.Subscribe.*
+import com.gu.productmove.zuora.model.InvoiceId
 import com.gu.productmove.zuora.rest.{ZuoraClientLive, ZuoraGetLive}
-import com.gu.productmove.zuora.{
-  GetSubscription,
-  InvoiceItemAdjustment,
-  InvoiceItemAdjustmentLive,
-  InvoiceItemWithTaxDetails,
-  TaxDetails,
-}
+import com.gu.productmove.zuora.*
 import zio.*
 import zio.json.JsonDecoder
 import zio.test.*
